elevated micros timer to 64-bit.  Should fix #89

+++ include/estimator.h
@@ -23,7 +23,7 @@ extern state_t _current_state;
 extern vector_t _adaptive_gyro_bias;
 
 void init_estimator(bool use_matrix_exponential, bool use_quadratic_integration, bool use_accelerometer);
-void run_estimator(uint32_t now);
+void run_estimator(uint64_t now);
 #ifdef __cplusplus
 }
 #endif
